fix: Still jumps back when realizing, but jumps back
I believe this is a fight between the TranslateTransform and the
ArrangeOverride.  Originally in ArrangeOverride I was attempting to
clear the TranslateTransform.  This didn't work though because in the
base class LoopItemsPanel it constancly sets the offsets back to the
OffsetFromInitialPosition.

What would happen if I explicitly set the offset of the newly added item
to the OffsetFromInitialPosition?